updated Android Common lib

+++ androidCommon/src/main/java/com/asksven/android/common/kernelutils/State.java
@@ -15,15 +15,17 @@
  */
 package com.asksven.android.common.kernelutils;
 
-import android.util.Log;
-
-import com.asksven.android.common.dto.StateDto;
-import com.asksven.android.common.privateapiproxies.StatElement;
-import com.google.gson.annotations.SerializedName;
-
 import java.io.Serializable;
 import java.text.DecimalFormat;
 import java.util.List;
+
+import android.util.Log;
+
+import com.asksven.android.common.dto.StateDto;
+import com.asksven.android.common.nameutils.UidInfo;
+import com.asksven.android.common.privateapiproxies.Process;
+import com.asksven.android.common.privateapiproxies.StatElement;
+import com.google.gson.annotations.SerializedName;
 
 
 /**
@@ -206,8 +208,8 @@ public class State extends StatElement implements Comparable<State>, Serializabl
 	public int compareTo(State o)
 	{
 		// we want to sort in descending order
-        return (o.m_freq) - (this.m_freq);
-    }
+		return ((int)( (o.m_freq) - (this.m_freq) ));
+	}
 
 
 }
